move project store types
